+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 LifterLMS Labs Changelog
 ========================
 
+= v1.2.0 - 2017-03-27 =
+-----------------------
+
+### Action Manager Lab
+
++ Check boxes to remove LifterLMS elements added to courses, lessons, and memberships instead of writing code
+
+### Lifti (Divi compatibility) Upgrades!
+
++ Divi Friends is now called Lifti because it's more fun
++ Add options for Divi page builder and custom layout support on LifterLMS courses, lessons, and memberships
++ Add a prebuilt layout for LifterLMS courses which automatically adds default LifterLMS content via LifterLMS shortcodes to a pagebuilder layout
+	+ If you already had this lab enabled, disable and re-enable so the layout will be installed!
++ Combine this lab with the Action Manager for a lot of Divi fun!
+
+
 = v1.1.2 - 2017-02-20 =
 -----------------------